Fix : 검색결과 페이지_결과 업데이트 안되는 것 수정
