+++ src/pages/search/SearchResultsPage.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useMemo } from 'react';
 import styled from 'styled-components';
-import { useNavigate } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 import SearchBox from './SearchComponents/SearchBox';
 import CategoryTab from './SearchResultsComponents/CategoryTab';
 import ResultsTotal from './SearchResultsComponents/ResultsTotal';
@@ -8,6 +8,10 @@ import ResultsCard from './SearchResultsComponents/ResultsCard';
 import ResultsLoan from './SearchResultsComponents/ResultsLoan';
 import ResultsSavings from './SearchResultsComponents/ResultsSavings';
 import ResultsSubscription from './SearchResultsComponents/ResultsSubscription';
+
+import { useDispatch } from 'react-redux';
+import { getSearchResults } from '../../common/api/Api';
+import { addCardResults, addLoanResults, addSavingsResults, addSubscriptionResults } from '../../store/searchSlice';
 
 const SearchResults = () => {
   // 자동 로그아웃
@@ -19,7 +23,31 @@ const SearchResults = () => {
     }
   });
 
-  // 탭 변경
+  // 검색결과
+  const dispatch = useDispatch();
+  const params = useParams();
+  const [keywordParams, setKeywordParams] = useState('');
+
+  console.log(keywordParams);
+
+  useEffect(() => {
+    params.keywords !== undefined && setKeywordParams(params.keywords);
+  }, [params]);
+
+  useEffect(() => {
+    const getServerResultData = async () => {
+      const cardData = await getSearchResults(keywordParams, 'card', 1);
+      dispatch(addCardResults(cardData));
+      const loanData = await getSearchResults(keywordParams, 'loan', 1);
+      dispatch(addLoanResults(loanData));
+      const savingsData = await getSearchResults(keywordParams, 'savings', 1);
+      dispatch(addSavingsResults(savingsData));
+      const subscriptionData = await getSearchResults(keywordParams, 'subscription', 1);
+      dispatch(addSubscriptionResults(subscriptionData));
+    };
+    keywordParams && getServerResultData();
+  }, [keywordParams]);
+
   const [tabIndex, setTabIndex] = useState(0);
   const [category, setCategory] = useState([
     { category: 'total', title: '통합', content: <ResultsTotal setTabIndex={setTabIndex} /> },
